move util to separate file
