+++ frontend/tests/e2e/search/search.spec.ts
@@ -1,4 +1,5 @@
-import { Locator, Page, expect, test } from "@playwright/test";
+import { clickAccordionWithTitle, clickMobileNavMenu, clickSearchNavLink, expectCheckboxIDIsChecked, expectSortBy, expectURLContainsQueryParam, fillSearchInputAndSubmit, getMobileMenuButton, getSearchInput, hasMobileMenu, refreshPageWithCurrentURL, selectSortBy, toggleCheckboxes, waitForSearchResultsLoaded } from "./searchUtil";
+import { expect, test } from "@playwright/test";
 
 test("should navigate from index to search page", async ({ page }) => {
   // Start from the index page with feature flag set
@@ -140,128 +141,3 @@ test.describe("Search page tests", () => {
     }
   });
 });
-
-// =========================
-// Test Helper Functions
-// =========================
-
-function getSearchInput(page: Page) {
-  return page.locator("#query");
-}
-
-async function fillSearchInputAndSubmit(term: string, page: Page) {
-  const searchInput = getSearchInput(page);
-  await searchInput.fill(term);
-  await page.click(".usa-search >> button[type='submit']");
-  expectURLContainsQueryParam(page, "query", term);
-}
-
-function expectURLContainsQueryParam(
-  page: Page,
-  queryParamName: string,
-  queryParamValue: string,
-) {
-  const currentURL = page.url();
-  expect(currentURL).toContain(`${queryParamName}=${queryParamValue}`);
-}
-
-async function waitForURLContainsQueryParam(
-  page: Page,
-  queryParamName: string,
-  queryParamValue: string,
-  timeout = 30000, // query params get set after a debounce period
-) {
-  const endTime = Date.now() + timeout;
-
-  while (Date.now() < endTime) {
-    const url = new URL(page.url());
-    const params = new URLSearchParams(url.search);
-    const actualValue = params.get(queryParamName);
-
-    if (actualValue === queryParamValue) {
-      return;
-    }
-
-    await page.waitForTimeout(500);
-  }
-
-  throw new Error(
-    `URL did not contain query parameter ${queryParamName}=${queryParamValue} within ${timeout}ms`,
-  );
-}
-
-async function clickSearchNavLink(page: Page) {
-  await page.click("nav >> text=Search");
-}
-
-function getMobileMenuButton(page: Page) {
-  return page.locator("button >> text=MENU");
-}
-
-async function hasMobileMenu(page: Page) {
-  const menuButton = getMobileMenuButton(page);
-  return await menuButton.isVisible();
-}
-
-async function clickMobileNavMenu(menuButton: Locator) {
-  await menuButton.click();
-}
-
-async function expectCheckboxIDIsChecked(page: Page, idWithHash: string) {
-  const checkbox: Locator = page.locator(idWithHash);
-  await expect(checkbox).toBeChecked();
-}
-
-async function toggleCheckboxes(
-  page: Page,
-  checkboxObject: Record<string, string>,
-  queryParamName: string,
-) {
-  let runningQueryParams = "";
-  for (const [checkboxID, queryParamValue] of Object.entries(checkboxObject)) {
-    await toggleCheckbox(page, checkboxID);
-    runningQueryParams += runningQueryParams
-      ? `,${queryParamValue}`
-      : queryParamValue;
-    await waitForURLContainsQueryParam(
-      page,
-      queryParamName,
-      runningQueryParams,
-    );
-  }
-}
-
-async function toggleCheckbox(page: Page, idWithoutHash: string) {
-  const checkBox = page.locator(`label[for=${idWithoutHash}]`);
-  await checkBox.isEnabled();
-  await checkBox.click();
-}
-
-async function refreshPageWithCurrentURL(page: Page) {
-  const currentURL = page.url();
-  await page.goto(currentURL); // go to new url in same tab
-  return page;
-}
-
-async function selectSortBy(page: Page, sortByValue: string) {
-  await page.locator("#search-sort-by-select").selectOption(sortByValue);
-}
-
-async function expectSortBy(page: Page, value: string) {
-  const selectedValue = await page
-    .locator('select[name="search-sort-by"]')
-    .inputValue();
-  expect(selectedValue).toBe(value);
-}
-
-async function waitForSearchResultsLoaded(page: Page) {
-  // Wait for number of opportunities to show
-  const resultsHeading = page.locator('h2:has-text("Opportunities")');
-  await resultsHeading.waitFor({ state: "visible", timeout: 60000 });
-}
-
-async function clickAccordionWithTitle(page: Page, accordionTitle: string) {
-  await page
-    .locator(`button.usa-accordion__button:has-text("${accordionTitle}")`)
-    .click();
-}
